+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+    - [cmd] added `set` command to change all editable issue meta fields (GitHub Issue #18)
     - [cmd] added `edit` command to edit issue descriptions (markdown) with external text editor (GitHub Issue #8)
     - [config] added `.gitissues/description.md` used as template when new issue is created (GitHub Issue #14)
     - [meta] added fields `type`, `labels` and `assignee` (GitHub Issue #4, #5, #10)
@@ -17,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
     - [structure] renamed `issue.md` to `description.md`
+
+### Removed
+
+    - [cmd] removed `state` command, replaced with `set` command (GitHub Issue #18)
 
 ## [v0.1.0] - 2025-12-21